refactor: store location
Change store folder location from src/store to src/app/store
